fix: if member is last, prevent destroy member

+++ app/controllers/api/v1/members_controller.rb
@@ -33,6 +33,12 @@ module Api
           response_not_found(Member.name)
           return
         end
+
+        if member.group.members.count <= 1
+          response_bad_request_with_custome_message(I18n.t('errors.format', attribute: I18n.t('activerecord.models.member'), message: I18n.t('errors.messages.last_member')))
+          return
+        end
+
         if member.destroy
           response_success(member)
         else
